Rename common to TypeScript
